requirements.txt
Flask==2.3.3
mediapipe==0.10.9
opencv-python==4.9.0.80
numpy==1.24.4
gunicorn>=21.2
psycopg2-binary>=2.9